+++ crates/alint-core/src/lib.rs
@@ -19,11 +19,12 @@ mod walker;
 pub mod when;
 
 pub use config::{
-    Config, ContentSourceSpec, ExtendsEntry, FileAppendFinalNewlineFixSpec, FileAppendFixSpec,
-    FileCollapseBlankLinesFixSpec, FileCreateFixSpec, FileNormalizeLineEndingsFixSpec,
-    FilePrependFixSpec, FileRemoveFixSpec, FileRenameFixSpec, FileStripBidiFixSpec,
-    FileStripBomFixSpec, FileStripZeroWidthFixSpec, FileTrimTrailingWhitespaceFixSpec, FixSpec,
-    NestedRuleSpec, PathsSpec, RuleSpec, resolve_content_source,
+    CompiledNestedSpec, Config, ContentSourceSpec, ExtendsEntry, FileAppendFinalNewlineFixSpec,
+    FileAppendFixSpec, FileCollapseBlankLinesFixSpec, FileCreateFixSpec,
+    FileNormalizeLineEndingsFixSpec, FilePrependFixSpec, FileRemoveFixSpec, FileRenameFixSpec,
+    FileStripBidiFixSpec, FileStripBomFixSpec, FileStripZeroWidthFixSpec,
+    FileTrimTrailingWhitespaceFixSpec, FixSpec, NestedRuleSpec, PathsSpec, RuleSpec,
+    resolve_content_source,
 };
 pub use engine::{Engine, RuleEntry};
 pub use error::{Error, Result};